Add jQuery Migrate v1.2.1
Migrate older jQuery code to jQuery 1.9+
This plugin can be used to detect and restore APIs or features that have
been deprecated in jQuery and removed as of version 1.9. This plugin can
be used to detect and restore APIs or features that have been deprecated
in jQuery and removed as of version 1.9.

https://github.com/jquery/jquery-migrate/